make common Link methods static and move to a separate class
